+++ pallets/forest/src/benchmarking.rs
@@ -42,24 +42,18 @@ mod benchmarks {
         let x25519_public_key = NULL_ARR;
         let endpoint = vec![];
 
-        let joining_server_info = JoiningForestServerInfo { x25519_public_key, endpoint };
-
-        let quote = prepare_attestation_for_validate::<T>(
+        let tdx_quote = prepare_attestation_for_validate::<T>(
             caller.clone(),
             x25519_public_key,
             QuoteContext::ForestAddTree,
         );
 
+        let server_info = ForestServerInfo { x25519_public_key, endpoint, tdx_quote };
+
         #[extrinsic_call]
-        _(RawOrigin::Signed(caller.clone()), joining_server_info.clone(), quote);
+        _(RawOrigin::Signed(caller.clone()), server_info.clone());
 
-        let tree_info = Trees::<T>::get(caller.clone()).unwrap();
-
-        let server_info = ForestServerInfo {
-            endpoint: joining_server_info.endpoint,
-            x25519_public_key: joining_server_info.x25519_public_key,
-            provisioning_certification_key: tree_info.provisioning_certification_key,
-        };
+        let _tree_info = Trees::<T>::get(caller.clone()).unwrap();
 
         assert_last_event::<T>(Event::<T>::TreeAdded { tree_account: caller, server_info }.into());
     }