+++ CHANGELOG.md
@@ -29,6 +29,7 @@ At the moment this project **does not** adhere to
 - Only read from the entropy-tss key-value store following a reboot ([#1429](https://github.com/entropyxyz/entropy-core/pull/1429))
 - Update substrate to polkadot stable2503-4 ([#1439](https://github.com/entropyxyz/entropy-core/pull/1439))
 - Rename outtie to forest ([#1481](https://github.com/entropyxyz/entropy-core/pull/1481))
+- Store full tdx quote in ForestServerInfo ([#1484](https://github.com/entropyxyz/entropy-core/pull/1484))
 
 ## [0.4.0](https://github.com/entropyxyz/entropy-core/compare/release/v0.3.0...release/v0.4.0) - 2025-03-31